requirements.txt
python-dotenv
openai
streamlit
pypdf
langchain_community
streamlit
faiss-cpu
cryptography
chromadb
langchain
tiktoken